+++ examples/scripts/comparison_examples/adamw.py
@@ -1,7 +1,6 @@
 import os
 
 import numpy as np
-import pybamm
 
 import pybop
 
@@ -12,9 +11,8 @@ dataset_path = os.path.join(
 )
 
 # Define model and use high-performant solver for sensitivities
-solver = pybamm.IDAKLUSolver()
 parameter_set = pybop.ParameterSet("Chen2020")
-model = pybop.lithium_ion.SPM(parameter_set=parameter_set, solver=solver)
+model = pybop.lithium_ion.SPM(parameter_set=parameter_set)
 
 # Fitting parameters
 parameters = pybop.Parameters(