+++ examples/scripts/getting_started/ask-tell-interface.py
@@ -1,7 +1,6 @@
 import os
 
 import numpy as np
-import pybamm
 
 import pybop
 
@@ -11,10 +10,9 @@ dataset_path = os.path.join(
     current_dir, "../../data/synthetic/spm_charge_discharge_75.csv"
 )
 
-# Define model and use high-performant solver for sensitivities
-solver = pybamm.IDAKLUSolver()
+# Define model
 parameter_set = pybop.ParameterSet.pybamm("Chen2020")
-model = pybop.lithium_ion.SPM(parameter_set=parameter_set, solver=solver)
+model = pybop.lithium_ion.SPM(parameter_set=parameter_set)
 
 # Fitting parameters
 parameters = pybop.Parameters(
@@ -67,7 +65,7 @@ x_best = []
 f_best = []
 
 # Run optimisation
-for i in range(100):
+for i in range(50):
     x = optim.optimiser.ask()
     f = [cost(x[0], calculate_grad=True)]
     optim.optimiser.tell(f)