+++ tests/integration/test_eis_parameterisation.py
@@ -128,12 +128,16 @@ class TestEISParameterisation:
         # Construct optimisation object
         common_args = {
             "cost": cost,
-            "max_iterations": 250,
+            "max_iterations": 100,
             "absolute_tolerance": 1e-6,
             "max_unchanged_iterations": 35,
             "sigma0": [0.05, 0.05, 1e-3]
             if isinstance(cost, pybop.GaussianLogLikelihood)
             else 0.02,
+            "polish": False
+            if isinstance(optimiser, pybop.SciPyDifferentialEvolution)
+            else None,
+            "population_size": 4,
         }
 
         # Create optimiser